refactor: 해당 dto 일관성 있게 리팩토링
