--- src/main/java/com/project/trainingdiary/dto/response/ExampleResponseDto.java
@@ -1,12 +1,0 @@
-package com.project.trainingdiary.dto.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class ExampleResponseDto {
-
-  private String name;
-  private int age;
-}
