+++ src/main/java/com/project/trainingdiary/model/ScheduleResponseDetail.java
@@ -1,8 +1,7 @@
 package com.project.trainingdiary.model;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
-import com.fasterxml.jackson.annotation.JsonInclude;
-import com.fasterxml.jackson.annotation.JsonInclude.Include;
+import com.project.trainingdiary.model.type.ScheduleStatusType;
 import java.time.LocalTime;
 import lombok.Builder;
 import lombok.Getter;
@@ -21,5 +20,5 @@ public class ScheduleResponseDetail {
   private String trainerName;
   private Long traineeId;
   private String traineeName;
-  private ScheduleStatus status;
+  private ScheduleStatusType status;
 }
